@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-class LighthouseSample < ActiveRecord::Base
+class LighthouseSample < ApplicationRecord
 end
