@@ -1,4 +1,4 @@
-class StockResource < ActiveRecord::Base
+class StockResource < ApplicationRecord
   include ResourceTools
   include CompositeResourceTools
 
@@ -10,9 +10,9 @@ class StockResource < ActiveRecord::Base
   json do
     translate(
       stock_resource_id: :id_stock_resource_lims,
-      uuid:              :stock_resource_uuid,
-      machine_barcode:   :labware_machine_barcode,
-      human_barcode:     :labware_human_barcode
+      uuid: :stock_resource_uuid,
+      machine_barcode: :labware_machine_barcode,
+      human_barcode: :labware_human_barcode
     )
 
     has_nested_model(:samples) do
