@@ -1,6 +1,6 @@
 class IgnoreableResource
   def self.create_or_update_from_json(*_args)
-    self.new
+    new
   end
 
   def id
