@@ -1,5 +1,5 @@
 class AddIndexToStockResources < ActiveRecord::Migration
   def change
-    add_index :stock_resource, [:id_stock_resource_lims, :id_sample_tmp, :id_lims], name: :composition_lookup_index
+    add_index :stock_resource, %i[id_stock_resource_lims id_sample_tmp id_lims], name: :composition_lookup_index
   end
 end
