@@ -19,10 +19,11 @@ module ForeignKeyConstraint
   def parse_options(table, modl, options)
     fk = options[:foreign_key] || 'id'
     as = options[:as] || "#{modl.singularize}_id"
-    raise 'Invalid table name' unless /\A[a-z_]+\Z/ === table
-    raise 'Invalid model name' unless /\A[a-z_]+\Z/ === modl
-    raise 'Invalid foreign key' unless /\A[a-z_]+\Z/ === fk
-    raise 'Invalid association' unless /\A[a-z_]+\Z/ === as
+    raise 'Invalid table name' unless /\A[a-z_]+\Z/.match?(table)
+    raise 'Invalid model name' unless /\A[a-z_]+\Z/.match?(modl)
+    raise 'Invalid foreign key' unless /\A[a-z_]+\Z/.match?(fk)
+    raise 'Invalid association' unless /\A[a-z_]+\Z/.match?(as)
+
     yield(table, modl, as, fk)
   end
 end
