@@ -8,6 +8,7 @@ module ResourceTools::CoreExtensions
       # [key, value]
       def convert(object)
         return object if object.is_a?(Array)
+
         [object]
       end
     end
@@ -17,7 +18,7 @@ module ResourceTools::CoreExtensions
     # Determines if this hash is within an acceptable bounds of the keys common with the
     # given hash.  It is assumed that values missing from 'other' are unchanged.
     def within_acceptable_bounds?(other)
-      (self.keys & other.keys).all? do |key|
+      (keys & other.keys).all? do |key|
         self[key].within_acceptable_bounds?(other[key])
       end
     end
@@ -38,14 +39,14 @@ module ResourceTools::CoreExtensions
   module String
     def within_acceptable_bounds?(value)
       return false if value.nil?
+
       self == value.to_s
     end
 
     def to_boolean_from_arguments
-      case
-      when ['true', 'yes'].include?(self.downcase) then true
-      when ['false', 'no'].include?(self.downcase) then false
-      else raise "Cannot convert #{self.inspect} to a boolean safely!"
+      if %w[true yes].include?(downcase) then true
+      elsif %w[false no].include?(downcase) then false
+      else raise "Cannot convert #{inspect} to a boolean safely!"
       end
     end
   end
@@ -65,6 +66,7 @@ module ResourceTools::CoreExtensions
 
     def within_acceptable_bounds?(v)
       return false if v.nil?
+
       (self - v).abs < numeric_tolerance
     end
   end
