@@ -8,11 +8,11 @@ describe QcResult do
     let(:expected_entries) { 1 }
     let(:json) do
       { 'id_qc_result_lims' => '123',
-        'assay'        => 'qPCR 1.0',
-        'value'        => '5.43',
-        'units'        => 'nM',
-        'cv'           => 2.34,
-        'qc_type'      => 'Molarity',
+        'assay' => 'qPCR 1.0',
+        'value' => '5.43',
+        'units' => 'nM',
+        'cv' => 2.34,
+        'qc_type' => 'Molarity',
         'id_pool_lims' => 'NT10369L',
         'labware_purpose' => 'Stock Plate',
         'date_created' => '2012-03-11 10:22:42',
@@ -34,11 +34,11 @@ describe QcResult do
     let(:expected_entries) { 0 }
     let(:json) do
       { 'id_qc_result_lims' => '123',
-        'assay'        => 'qPCR 1.0',
-        'value'        => '5.43',
-        'units'        => 'nM',
-        'cv'           => 2.34,
-        'qc_type'      => 'Molarity',
+        'assay' => 'qPCR 1.0',
+        'value' => '5.43',
+        'units' => 'nM',
+        'cv' => 2.34,
+        'qc_type' => 'Molarity',
         'id_pool_lims' => 'NT10369L',
         'labware_purpose' => 'Stock Plate',
         'date_created' => '2012-03-11 10:22:42',
@@ -57,7 +57,7 @@ describe QcResult do
     it 'has an index on id_qc_result_lims and :id_lims' do
       expect(
         ActiveRecord::Base.connection.index_exists?(
-          :qc_result, [:id_qc_result_lims, :id_lims], name: 'lookup_index'
+          :qc_result, %i[id_qc_result_lims id_lims], name: 'lookup_index'
         )
       ).to be_truthy
     end
