@@ -12,12 +12,11 @@ describe StockResource do
     it_behaves_like 'maps JSON fields', {
     }
 
-    it_behaves_like 'ignores JSON fields', [
-    ]
+    it_behaves_like 'ignores JSON fields', []
 
-    it_behaves_like 'belongs to', [
-      :sample,
-      :study
+    it_behaves_like 'belongs to', %i[
+      sample
+      study
     ], [:samples]
 
     it_behaves_like 'a nested resource'
@@ -31,32 +30,32 @@ describe StockResource do
 
     let(:json) do
       {
-        "created_at" => "2012-03-11 10:22:42",
-        "updated_at" => "2012-03-11 10:22:42",
+        'created_at' => '2012-03-11 10:22:42',
+        'updated_at' => '2012-03-11 10:22:42',
 
-        "samples" => [
-          "sample_uuid" => "000000-0000-0000-0000-0000000000",
-          "study_uuid" => "000000-0000-0000-0000-0000000001"
+        'samples' => [
+          'sample_uuid' => '000000-0000-0000-0000-0000000000',
+          'study_uuid' => '000000-0000-0000-0000-0000000001'
         ],
 
-        "stock_resource_id" => "12345",
-        "stock_resource_uuid" => "000000-0000-0000-0000-0000000002",
+        'stock_resource_id' => '12345',
+        'stock_resource_uuid' => '000000-0000-0000-0000-0000000002',
 
         # We don't validate the barcode format, other
         # LIMS might use different standards
-        "labware_machine_barcode" => "1220456987123",
-        "labware_human_barcode" => "DN456987D",
-        "labware_coordinate" => "A1",
-        "labware_type" => "well",
+        'labware_machine_barcode' => '1220456987123',
+        'labware_human_barcode' => 'DN456987D',
+        'labware_coordinate' => 'A1',
+        'labware_type' => 'well',
 
-        "current_volume" => 1.23,
-        "initial_volume" => 4.56,
-        "concentration"  => 23.56,
+        'current_volume' => 1.23,
+        'initial_volume' => 4.56,
+        'concentration' => 23.56,
 
-        "gel_pass" => "Pass",
-        "pico_pass" => "Pass",
-        "snp_count" => 2,
-        "measured_gender" => "Unknown"
+        'gel_pass' => 'Pass',
+        'pico_pass' => 'Pass',
+        'snp_count' => 2,
+        'measured_gender' => 'Unknown'
 
       }
     end
@@ -67,7 +66,7 @@ describe StockResource do
       let(:updated_json) do
         updated_json = json.dup
         updated_json['current_volume'] = 1.00
-        updated_json['updated_at'] = "2012-03-11 10:22:42"
+        updated_json['updated_at'] = '2012-03-11 10:22:42'
         updated_json
       end
 
@@ -89,32 +88,32 @@ describe StockResource do
 
     let(:json) do
       {
-        "created_at" => "2012-03-11T10:22:42+00:00",
-        "updated_at" => "2012-03-11T10:22:42+00:00",
+        'created_at' => '2012-03-11T10:22:42+00:00',
+        'updated_at' => '2012-03-11T10:22:42+00:00',
 
-        "samples" => [
-          "sample_uuid" => "000000-0000-0000-0000-0000000000",
-          "study_uuid" => "000000-0000-0000-0000-0000000001"
+        'samples' => [
+          'sample_uuid' => '000000-0000-0000-0000-0000000000',
+          'study_uuid' => '000000-0000-0000-0000-0000000001'
         ],
 
-        "stock_resource_id" => 12345,
-        "stock_resource_uuid" => "000000-0000-0000-0000-0000000002",
+        'stock_resource_id' => 12_345,
+        'stock_resource_uuid' => '000000-0000-0000-0000-0000000002',
 
         # We don't validate the barcode format, other
         # LIMS might use different standards
-        "machine_barcode" => "1220456987123",
-        "human_barcode" => "DN456987D",
-        "labware_coordinate" => "A1",
-        "labware_type" => "well",
+        'machine_barcode' => '1220456987123',
+        'human_barcode' => 'DN456987D',
+        'labware_coordinate' => 'A1',
+        'labware_type' => 'well',
 
-        "current_volume" => 1.23,
-        "initial_volume" => 4.56,
-        "concentration"  => 23.56,
+        'current_volume' => 1.23,
+        'initial_volume' => 4.56,
+        'concentration' => 23.56,
 
-        "gel_pass" => "Pass",
-        "pico_pass" => "Pass",
-        "snp_count" => 2,
-        "measured_gender" => "Unknown"
+        'gel_pass' => 'Pass',
+        'pico_pass' => 'Pass',
+        'snp_count' => 2,
+        'measured_gender' => 'Unknown'
 
       }
     end
@@ -125,22 +124,22 @@ describe StockResource do
 
     let(:json) do
       {
-        "created_at" => "2012-03-11 10:22:42",
-        "updated_at" => "2012-03-11 10:22:42",
+        'created_at' => '2012-03-11 10:22:42',
+        'updated_at' => '2012-03-11 10:22:42',
 
-        "samples" => [
-          "sample_uuid" => "000000-0000-0000-0000-0000000000",
-          "study_uuid" => "000000-0000-0000-0000-0000000001"
+        'samples' => [
+          'sample_uuid' => '000000-0000-0000-0000-0000000000',
+          'study_uuid' => '000000-0000-0000-0000-0000000001'
         ],
 
-        "stock_resource_id" => "12345",
-        "stock_resource_uuid" => "000000-0000-0000-0000-0000000002",
+        'stock_resource_id' => '12345',
+        'stock_resource_uuid' => '000000-0000-0000-0000-0000000002',
 
         # We don't validate the barcode format, other
         # LIMS might use different standards
-        "machine_barcode" => "3040456987123",
-        "human_barcode" => "NT456987D",
-        "labware_type" => "tube"
+        'machine_barcode' => '3040456987123',
+        'human_barcode' => 'NT456987D',
+        'labware_type' => 'tube'
 
         # QC data will not always be present, especially for tubes
         # So lets make sure its absence doesn't cause issues
